requirements.txt
openai==0.27.1
python-dotenv==0.21.0
llama-index